rename InstanceCache to ResourceManager and move file location:
